api finished

+++ estacionamento.Application/Dtos/EstabelecimentoDto.cs
@@ -6,7 +6,7 @@
 
         public string Nome { get; set; }
 
-        public int CNPJ { get; set; }
+        public long CNPJ { get; set; }
 
         public string Endereco { get; set; }
 
@@ -14,7 +14,5 @@
         public int VagaCarro { get; set; }
 
         public int VagaMoto { get; set; }
-
-
     }
 }
